# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a currency exchange API.",
        "title": "Currency exchange API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/convert": {
            "get": {
                "description": "Get the exchange rate between two currencies with a given amount",
                "produces": [
                    "application/json"
                ],
                "summary": "Get amount converted from one currency to another",
                "operationId": "get-exchange-rate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "From currency code",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To currency code",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Amount to convert",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}